refact(wip): move method to formService and utils
