+++ frontend/app/services/form.service.ts
@@ -6,7 +6,7 @@ import { ISite, ISitesGroup } from '../interfaces/geom';
 import { JsonData } from '../types/jsondata';
 import { Utils } from '../utils/utils';
 import { MonitoringObjectService } from './monitoring-object.service';
-import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
+import { FormBuilder, FormControl, FormGroup, FormArray, AbstractControl } from '@angular/forms';
 import { IExtraForm, IFormMap } from '../interfaces/object';
 import { ConfigService } from './config.service';
 
@@ -176,5 +176,143 @@ export class FormService {
       // }
     }
     return postData;
+  }
+
+  /**
+   * Add a form control to the object form.
+   *
+   * @param {Object} frmCtrl - The form control to add to the object form
+   * @param {FormGroup} objForm - The object form to add the form control to
+   * @return {Observable<FormGroup>} The updated object form
+   */
+  addFormCtrlToObjForm(
+    frmCtrl: { frmCtrl: FormControl; frmName: string },
+    objForm: FormGroup
+  ): Observable<FormGroup> {
+    if (frmCtrl.frmName in objForm.controls) {
+      // Si le champ existe déjà dans l'objet form, on ne fait rien
+    } else {
+      objForm.addControl(frmCtrl.frmName, frmCtrl.frmCtrl);
+    }
+    return of(objForm);
+  }
+
+  /**
+   * Add multiple form groups to the object form.
+   *
+   * @param {Object} formGroups - Object containing form groups to add
+   * @param {FormGroup} targetForm - The target form group to add form groups to
+   * @return {FormGroup} The updated target form group
+   */
+  addMultipleFormGroupsToObjForm(
+    formGroups: { [key: string]: FormGroup },
+    targetForm: FormGroup
+  ): FormGroup {
+    let dynamicGroups = targetForm.get('dynamicGroups') as FormArray;
+
+    if (!dynamicGroups) {
+      dynamicGroups = this._formBuilder.array([]);
+      targetForm.addControl('dynamicGroups', dynamicGroups);
+      dynamicGroups = targetForm.get('dynamicGroups') as FormArray; // Refresh reference after adding it
+    }
+
+    for (let i = dynamicGroups.controls.length - 1; i >= 0; i--) {
+      const control = dynamicGroups.controls[i];
+      const controlName = control.get('name')?.value;
+      if (!formGroups[controlName]) {
+        dynamicGroups.removeAt(i);
+      }
+    }
+
+    for (const key in formGroups) {
+      const existingControlIndex = dynamicGroups.controls.findIndex(
+        (control) => control.get('name')?.value === key
+      );
+
+      if (existingControlIndex !== -1) {
+        dynamicGroups.controls[existingControlIndex].patchValue(formGroups[key].value, {
+          emitEvent: false,
+        });
+      } else {
+        const newControl = formGroups[key];
+        newControl.addControl('name', this._formBuilder.control(key)); // Adding control with key as 'name'
+        dynamicGroups.push(newControl);
+      }
+    }
+    return targetForm;
+  }
+
+  /**
+   * Patches values inside dynamic form groups
+   *
+   * @param valuesToPatch Values to patch, with the key being the control name and the value
+   * being the new value to set
+   * @param objOfFormGroups Object containing form groups to patch, with the key being the group name
+   * and the value being the form group
+   */
+  patchValuesInDynamicGroups(
+    valuesToPatch: { [controlName: string]: any },
+    objOfFormGroups: { [groupName: string]: FormGroup }
+  ): void {
+    Object.keys(objOfFormGroups).forEach((groupName) => {
+      const formGroup = objOfFormGroups[groupName];
+      if (formGroup instanceof FormGroup) {
+        this.patchValuesInFormGroup(formGroup, valuesToPatch);
+      }
+    });
+  }
+
+  /**
+   * Patches values inside a form group
+   *
+   * @param formGroup Form group to patch values in
+   * @param valuesToPatch Values to patch, with the key being the control name and the value
+   * being the new value to set
+   */
+  patchValuesInFormGroup(formGroup: FormGroup, valuesToPatch: JsonData): void {
+    Object.keys(valuesToPatch).forEach((controlName) => {
+      if (formGroup.contains(controlName)) {
+        formGroup.get(controlName).patchValue(valuesToPatch[controlName]);
+      }
+    });
+  }
+
+  /**
+   * Flattens a FormGroup into a JSON object.
+   * @param formGroup The FormGroup to flatten.
+   * @returns The flattened FormGroup as a JSON object.
+   */
+  flattenFormGroup(formGroup: FormGroup): JsonData {
+    const flatObject: JsonData = {}; // JSON object to return
+
+    // Recursive function to process nested controls
+    /**
+     * Flattens a control (or controls within a FormGroup or FormArray) into the
+     * flattened JSON object.
+     * @param control The control (or FormGroup or FormArray) to flatten.
+     * @param keyPrefix The prefix of the control's key to use in the flattened
+     * object.
+     */
+    const flattenControl = (control: AbstractControl, keyPrefix: string = ''): void => {
+      if (control instanceof FormGroup) {
+        // If control is a FormGroup, recurse into its controls and flatten each one
+        Object.entries(control.controls).forEach(([controlName, nestedControl]) => {
+          flattenControl(nestedControl, `${controlName}.`);
+        });
+      } else if (control instanceof FormArray) {
+        // If control is a FormArray, recurse into each control in the FormArray
+        control.controls.forEach((arrayControl, index) => {
+          flattenControl(arrayControl, `${keyPrefix}`);
+        });
+      } else {
+        // If control is not a FormGroup or FormArray, add it to the flattened object
+        flatObject[keyPrefix.slice(0, -1)] = control.value;
+      }
+    };
+
+    // Start flattening from the root FormGroup
+    flattenControl(formGroup);
+
+    return flatObject;
   }
 }
